+++ app/Models/Profile.php
@@ -2,14 +2,14 @@
 
 namespace Uccello\Core\Models;
 
+use App\Models\UccelloModel;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Support\Collection;
 use Spatie\Searchable\Searchable;
 use Spatie\Searchable\SearchResult;
-use Uccello\Core\Database\Eloquent\Model;
 use Uccello\Core\Support\Traits\UccelloModule;
 
-class Profile extends Model implements Searchable
+class Profile extends UccelloModel implements Searchable
 {
     use SoftDeletes;
     use UccelloModule;